util/scripts/sync_tas/requirements.txt
git+https://github.com/ChameleonCloud/pytas#egg=pytas
mysql-connector-python
pytz
